backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
groq==0.4.1
python-dotenv==1.0.0
pymysql==1.1.0
psycopg2-binary==2.9.9
pandas==2.1.4
gunicorn==21.2.0
colorama==0.4.6
